use `uv` for env management, update Python versions
This switches to the `uv` package manager, which simplifies venv management.

In order to install the expected dev dependencies for the entire supported range
of Python versions, the minimum supported Python version had to be upped from
3.6 to 3.8. This is not a compatibility-breaking change, as downstream users can
still install the old version.

The Noxfile was updated to use `uv` as the package installation backend.

The GH Action was updated to install and use UV. I tested the new action with
the `act` tool.